+++ ML-Frameworks/pytorch-aarch64/build.sh
@@ -24,13 +24,32 @@ build_log=build-$(git rev-parse --short=7 HEAD)-$(date '+%Y-%m-%dT%H-%M-%S').log
 exec &> >(tee -a $build_log)
 
 # Bail out if sources are already there
-if [ -d pytorch ] || [ -d builder ] || [ -d ComputeLibrary ] ; then
-    echo "You appear to have sources already (pytorch/builder/ComputeLibrary)" \
+if [ -f .torch_build_container_id ] || [ -f .torch_ao_build_container_id ] || \
+    [ -d ao ] || [ -d ComputeLibrary ] || [ -d pytorch ]; then
+    printf "\n\n%s\n%s\n%s\n%s\n%s\n%s\n%s\n\n\n" \
+        "You appear to have artefacts from a previous build lying around." \
+        "Check for any of the following:" \
+        "  - .torch_build_container_id" \
+        "  - .torch_ao_build_container_id" \
+        "  - ao" \
+        "  - ComputeLibrary" \
+        "  - pytorch"
 
-    if ! ([[ $* == *--force* ]] || [[ $* == *--use-existing-sources* ]]) ; then
-        >&2 echo "rerun with --force to overwrite sources or with" \
-                 "--use-existing-sources to build your existing sources."
+    if [[ "$*" != *--fresh* ]] && [[ "$*" != *--use-existing-sources* ]]; then
+        >&2 printf "\n\n%s\n%s\n%s\n\n\n" \
+            "Rerun with one of the following options:" \
+            "  - '--fresh': wipe the pre-existing sources and do a fresh build" \
+            "  - '--use-existing-sources': reuse the sources as is"
         exit 1
+    fi
+
+    # Wipe old build artefacts
+    if [[ $* == *--fresh* ]]; then
+        if [ -f .torch_build_container_id ]; then rm -f .torch_build_container_id; fi
+        if [ -f .torch_ao_build_container_id ]; then rm -f .torch_ao_build_container_id; fi
+        if [ -d ao ]; then rm -rf ao; fi
+        if [ -d ComputeLibrary ]; then rm -rf ComputeLibrary; fi
+        if [ -d pytorch ]; then rm -rf pytorch; fi
     fi
 fi