+++ ML-Frameworks/pytorch-aarch64/CHANGELOG.md
@@ -19,6 +19,7 @@ where `YY` is the year, and `MM` the month of the increment.
   - TORCH_AO_HASH to 8e2ca35ea603349e71c2467e10fd371e34bf52bc, from main, September 23rd.
   - KLEIDIAI_HASH to bd2e6ae060014035e25bf4986be682762c446c2d, v1.14 from main.
 - Update torchvision from 0.23.0 to a nightly build, 0.25.0.dev20250923
+- Change of flag name in `./build.sh` from `--force` to `--fresh`
 
 ### Removed
 - Removes WIP ComputeLibrary patch https://review.mlplatform.org/c/ml/ComputeLibrary/+/12818/1.